heap: switch mem ops from 64 bit to 32 bit

+++ common/heap.h
@@ -5,8 +5,8 @@
 #define KHEAP_SIZE (8 * 1024 * 1024)
 
 void* kmalloc(uint64 size);
-void* memset(void* buf, char c, uint64 n);
-void* memcpy(void* dst, const void* src, uint64 n);
+void* memset(void* buf, char c, uint32 n);
+void* memcpy(void* dst, const void* src, uint32 n);
 
 #endif
 #ifdef IMPL_KMALLOC_1
@@ -30,7 +30,7 @@ void* kmalloc(uint64 size)
     return ptr;
 }
 
-void* memset(void* buf, char c, uint64 n)
+void* memset(void* buf, char c, uint32 n)
 {
     uint8* p = (uint8*)buf;
     while (n--)
@@ -38,7 +38,7 @@ void* memset(void* buf, char c, uint64 n)
     return buf;
 }
 
-void* memcpy(void* dst, const void* src, uint64 n)
+void* memcpy(void* dst, const void* src, uint32 n)
 {
     uint8* d       = (uint8*)dst;
     const uint8* s = (const uint8*)src;
